move AgentLogList to frontend and fix firebase imports
